issue/JC-501_update_docs_publishing_to_the_new_process
We now push the docs up to `docs.atlassian.com` via a separate script
(defined in the plan template). However, to generate docs with proper
links between modules, calling `mvn site` at the top-level isn't enough.
You need to call site-deploy (or site:stage).
And, in order to either site-deploy or site:stage, a url for deployment
has to exist. ¯\_(ツ)_/¯

So, `site-deploy` (invoked on `release:perform`) now dumps the full site
into the local user's directory: ~/jrjc-local-site-deployment.
The follow-on script in plan-templates tarballs it & takes care of
publishing it to docs.atlassian.com the new way.